separated routes into different controllers
