+++ app/Http/Controllers/HomeController.php
@@ -1,0 +1,52 @@
+<?php
+
+namespace PRStats\Http\Controllers;
+
+use Carbon\Carbon;
+use PRStats\Models\Player;
+
+class HomeController extends Controller
+{
+    public function index()
+    {
+        //top players
+        $players = Player::with('clan')
+            ->withCount(['matches' => function ($q) {
+                return $q->where('match_player.updated_at', '>=', Carbon::now()->startOfMonth());
+            }])
+            ->where('updated_at', '>=', Carbon::now()->startOfMonth())
+            ->orderBy('monthly_score', 'desc')
+            ->paginate(50);
+
+        $newest = Player::with('clan')
+            ->orderBy('created_at', 'desc')
+            ->limit(10)
+            ->get();
+
+        $longest = Player::with(['clan', 'matches'])
+            ->orderBy('minutes_played', 'desc')
+            ->limit(10)
+            ->get()
+            ->sortByDesc(function($player) {
+                return $player->minutesPlayed();
+            });
+
+        $mostKills = Player::with('clan')
+            ->orderBy('total_kills', 'desc')
+            ->limit(10)
+            ->get();
+
+        $mostDeaths = Player::with('clan')
+            ->orderBy('total_deaths', 'desc')
+            ->limit(10)
+            ->get();
+
+        return view('prstats.home', [
+            'players'    => $players,
+            'newest'     => $newest,
+            'longest'    => $longest,
+            'mostKills'  => $mostKills,
+            'mostDeaths' => $mostDeaths,
+        ]);
+    }
+}
