+++ app/Http/routes.php
@@ -1,29 +1,16 @@
 <?php
 
-/*
-|--------------------------------------------------------------------------
-| Application Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register all of the routes for an application.
-| It's a breeze. Simply tell Laravel the URIs it should respond to
-| and give it the controller to call when that URI is requested.
-|
-*/
+Route::get('/', 'HomeController@index');
 
-Route::get('/', 'Home@index');
+Route::get('clans', ['as' => 'clans', 'uses' => 'ClanController@index']);
+Route::post('clans', ['as' => 'clans.search', 'uses' => 'ClanController@search']);
+Route::get('clan/{id}/{slug}', ['as' => 'clan', 'uses' => 'ClanController@show']);
 
+Route::get('servers', ['as' => 'servers', 'uses' => 'ServerController@index']);
+Route::get('server/{id}/{slug}', ['as' => 'server', 'uses' => 'ServerController@show']);
+Route::get('match/{id}/{map}', ['as' => 'match', 'uses' => 'ServerController@match']);
 
-Route::get('clan/{id}/{slug}', ['as' => 'clan', 'uses' => 'Home@clan']);
-Route::get('player/{pid}/{slug}', ['as' => 'player', 'uses' => 'Home@player']);
-Route::get('server/{id}/{slug}', ['as' => 'server', 'uses' => 'Home@server']);
-
-Route::get('servers', ['as' => 'servers', 'uses' => 'Home@servers']);
-Route::get('players', ['as' => 'players', 'uses' => 'Home@players']);
-Route::get('clans', ['as' => 'clans', 'uses' => 'Home@clans']);
-Route::get('search', ['as' => 'players.search', 'uses' => 'Home@playerSearch']);
-Route::post('clans', ['as' => 'clans.search', 'uses' => 'Home@clanSearch']);
-Route::get('match/{id}/{map}', ['as' => 'match', 'uses' => 'Home@matchDetails']);
-Route::get('{slug}', ['as' => 'player.short', 'uses' => 'Home@playerShortUrl']);
-
-
+Route::get('players', ['as' => 'players', 'uses' => 'PlayerController@index']);
+Route::get('player/{pid}/{slug}', ['as' => 'player', 'uses' => 'PlayerController@show']);
+Route::get('search', ['as' => 'players.search', 'uses' => 'PlayerController@search']);
+Route::get('{slug}', ['as' => 'player.short', 'uses' => 'PlayerController@shortUrl']);
